requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=5.4.0
psutil>=5.8.0
requests>=2.26.0
